Fixing migrations and stuff

+++ db/migrate/20120123004161_migrate_data_and_photos.rb
@@ -11,13 +11,21 @@ class MigrateDataAndPhotos < ActiveRecord::Migration
 		  #Images sit directly in this directory
 		  files = ftp.chdir('people_images')
 		  Person.where("`pic` IS NOT NULL").all.first(5).each do |person|
-		  	ftp.getbinaryfile(person.pic, "tmp/" + person.pic)
-		  	puts "tmp/" + person.pic
-		  	person.picture = File.open("#{Rails.root}/tmp/#{person.pic}")
-		  	person.save!
-		  	File.delete("#{Rails.root}/tmp/#{person.pic}")
-		  end
-		  #remove_column :people, :pic
+		  	begin
+			  	ftp.getbinaryfile(person.pic, "tmp/" + person.pic)
+			  	
+			  	File.open("#{Rails.root}/tmp/#{person.pic}") do |file|
+			  		person.picture = file
+			  		person.save!
+			  	end
+			 	rescue
+			 		puts "unable to transfer person #{person.id} -> #{person.pic}"
+				ensure
+		  		File.delete("#{Rails.root}/tmp/#{person.pic}") if File.exist?("#{Rails.root}/tmp/#{person.pic}")
+		  	end
+	  	end
+		  remove_column :people, :pic
+
 		  #reset directory
 		  files = ftp.chdir('../')
 		  files = ftp.chdir('show_images')
@@ -26,6 +34,7 @@ class MigrateDataAndPhotos < ActiveRecord::Migration
 		  valid_dirs = ftp.nlst.map(&:to_i)
 		  
 		  Show.where(:id => valid_dirs).first(2).each do |show|
+		  	
 		  	ftp.chdir("#{show.id}")
 		  	filelist = ftp.nlst
 		  	
@@ -42,15 +51,22 @@ class MigrateDataAndPhotos < ActiveRecord::Migration
 		  			
 		  			if ts_range.include? file_ts
 		  				# Got it. Lez go
-		  				ftp.getbinaryfile(filename, "tmp/" + filename)
-					  	puts "Poster: tmp/" + filename
-					  	show.poster = File.open("#{Rails.root}/tmp/#{filename}")
-					  	show.save!
-					  	filelist = filelist - [filename]
+		  				begin
+			  				ftp.getbinaryfile(filename, "tmp/" + filename)
+						  	
+						  	File.open("#{Rails.root}/tmp/#{filename}") do |file|
+							  	show.poster = file
+							  	show.save!
+							  end
+						  	filelist = filelist - [filename]
+						  rescue
+						  	puts "unable to transfer poster #{show.id} -> #{filename}"
+						  ensure
+						  		File.delete("#{Rails.root}/tmp/#{filename}") if File.exist?("#{Rails.root}/tmp/#{filename}")
+						  end
 					  	break
 		  			end
 		  		end
-		  		  	
 		  	end
 		  			  	
 		  	# Take the rest, we know general format for old thumbs and stuff...dump those
@@ -61,12 +77,18 @@ class MigrateDataAndPhotos < ActiveRecord::Migration
 		  		
 		  		# move it over to S3
 		  		# upload a file
-		  		ftp.getbinaryfile(filename, "tmp/" + filename)
-			  	puts "tmp/" + filename
-			  	
-					basename = "shows/#{show.id}/misc/#{filename}"
-					o = s3_bucket.objects[basename]
-					o.write(:file => "#{Rails.root}/tmp/#{filename}")
+		  		begin
+			  		ftp.getbinaryfile(filename, "tmp/" + filename)
+				  	
+				  	
+						basename = "shows/#{show.id}/misc/#{filename}"
+						o = s3_bucket.objects[basename]
+						o.write(:file => "#{Rails.root}/tmp/#{filename}")
+					rescue
+						puts "unable to write file for show #{show.id} -> #{filename}"
+					ensure
+						File.delete("#{Rails.root}/tmp/#{filename}") if File.exist?("#{Rails.root}/tmp/#{filename}")
+					end
 		  	end
 		  	
 		  	# Switch back so we're ready for next show
